Quantiitative Finance Projects/volatility-pipeline copy/requirements.txt
yfinance
pandas
numpy
scipy
matplotlib
sqlalchemy
pymysql
schedule
plotly
streamlit
alpaca-py